backend/requirements.txt
pdfminer.six
transformers
networkx
pyvis
torch
Flask
flask_cors
PyExecJS
pandas
scikit-learn
PyMuPDF
